requirements.txt
aiohttp
asyncio
psycopg2-binary
python-dotenv